multiplayer-ai-chat/voice-agent/pyproject.toml
[project]
name = "nomi-voice-agent"
version = "0.1.0"
description = "LiveKit voice agent for multiplayer AI chat"
requires-python = ">=3.10"
dependencies = [
    "livekit>=0.17.0",
    "livekit-agents>=0.11.0",
    "livekit-plugins-openai>=0.8.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
